refactor trainer optimizer modules
