+++ src/prime_rl/trainer/optim/cpu_adam/kernel.cpp
