+++ src/prime_rl/trainer/optim/offload.py
@@ -1,3 +1,5 @@
+# This file is fully AI-generated.
+
 import copy
 import queue
 import threading
@@ -8,17 +10,15 @@ from typing import Callable
 
 import torch
 import torch.distributed as dist
-from dion import Muon
 from torch import nn
 from torch.distributed.tensor import DTensor
-from torch.optim import SGD, AdamW, Optimizer
+from torch.optim import AdamW, Optimizer
 
-from prime_rl.configs.trainer import OptimizerConfig, OptimizerOffloadingConfig
-from prime_rl.trainer.cpu_adam import adamw_step as native_cpu_adamw_step
-from prime_rl.trainer.cpu_adam import copy_or_add_bfloat16_multi_ as native_copy_or_add_bfloat16_multi_
-from prime_rl.trainer.cpu_adam import load_cpu_adamw_kernel
-from prime_rl.trainer.parallel_dims import ParallelDims
-from prime_rl.trainer.sign_sgd import SignSGD
+from prime_rl.configs.trainer import OptimizerOffloadingConfig
+from prime_rl.trainer.optim.cpu_adam import adamw_step as native_cpu_adamw_step
+from prime_rl.trainer.optim.cpu_adam import add_bfloat16_ as native_add_bfloat16_
+from prime_rl.trainer.optim.cpu_adam import copy_or_add_bfloat16_multi_ as native_copy_or_add_bfloat16_multi_
+from prime_rl.trainer.optim.cpu_adam import load_cpu_adamw_kernel
 from prime_rl.utils.logger import get_logger
 
 
@@ -1380,58 +1380,6 @@ class CPUOffloadOptimizer:
         self._initialized = True
 
 
-def setup_optimizer(
-    config: OptimizerConfig,
-    named_params: list[tuple[str, nn.Parameter]],
-    parallel_dims: ParallelDims,
-    lora: bool = False,
-    offload_config: OptimizerOffloadingConfig | None = None,
-    model: nn.Module | None = None,
-) -> tuple[Optimizer | CPUOffloadOptimizer, GradientOffloadManager | None]:
-    if offload_config is not None and config.type == "muon":
-        raise ValueError("CPU optimizer offload does not support Muon because the optimizer step runs on CPU")
-    if offload_config is not None and config.max_norm is not None:
-        get_logger().warning("Disabling gradient clipping because CPU optimizer offload updates during backward")
-        config.max_norm = None
-    if lora:
-        # Wait for run 0 to be created in the multi run manager
-        # Otherwise, the creation will reset the parameters
-        multi_run_manager = get_multi_run_manager()
-        multi_run_manager.wait_for_run(0)
-        named_params = multi_run_manager.get_named_parameters_for_run(0)
-
-    optimizer_named_params = named_params
-    master_weights = None
-    if offload_config is not None:
-        if model is None:
-            raise ValueError("CPU optimizer offload requires the model")
-        optimizer_named_params, master_weights = _create_cpu_master_weights(
-            model,
-            named_params,
-            pin_memory=not (config.type == "adamw" and offload_config.cpu_optimizer_backend == "native"),
-        )
-
-    optimizer = _create_optimizer(
-        config,
-        optimizer_named_params,
-        parallel_dims,
-        fused_adamw=offload_config is not None and config.type == "adamw",
-    )
-
-    if offload_config is not None:
-        assert master_weights is not None
-        get_logger().info("Using CPU offload for gradients and the optimizer step")
-        optimizer = CPUOffloadOptimizer(
-            optimizer,
-            offload_config=offload_config,
-            master_weights=master_weights,
-            dp_replicate=parallel_dims.dp_replicate,
-        )
-        return optimizer, optimizer._gradient_manager
-
-    return optimizer, None
-
-
 @torch.no_grad()
 def _create_cpu_master_weights(
     model: nn.Module,
@@ -1477,128 +1425,3 @@ def _create_cpu_master_weights(
         f"{'pinned' if pin_memory else 'pageable'} FP32 masters; persistent GPU parameters use BF16"
     )
     return master_named_params, master_weights
-
-
-def _create_optimizer(
-    config: OptimizerConfig,
-    named_params: list[tuple[str, nn.Parameter]],
-    parallel_dims: ParallelDims,
-    lr: float | None = None,
-    fused_adamw: bool = False,
-) -> Optimizer:
-    """Create optimizer. If lr is None, uses config.lr."""
-    if lr is None:
-        lr = config.lr
-    # Only hand trainable params to the optimizer. Frozen params (e.g. the DSA sparse
-    # indexer, which runs under no_grad) carry no optimizer state, and including them
-    # breaks strict checkpoint resume (DCP materializes state for every requires_grad
-    # param at load time, mismatching the saved state). Muon filters internally below.
-    trainable_params = [p for _, p in named_params if p.requires_grad]
-    match config.type:
-        case "sgd":
-            return SGD(
-                params=trainable_params,
-                lr=lr,
-                weight_decay=config.weight_decay,
-                momentum=config.momentum,
-                nesterov=config.nesterov,
-            )
-        case "adamw":
-            return AdamW(
-                params=trainable_params,
-                lr=lr,
-                weight_decay=config.weight_decay,
-                betas=(config.betas1, config.betas2),
-                fused=fused_adamw,
-            )
-        case "muon":
-            return _create_muon_optimizer(config, named_params, parallel_dims, lr)
-        case "sign_sgd":
-            return SignSGD(
-                params=trainable_params,
-                lr=lr,
-                weight_decay=config.weight_decay,
-            )
-
-
-def _create_muon_optimizer(
-    config: OptimizerConfig,
-    named_params: list[tuple[str, nn.Parameter]],
-    parallel_dims: ParallelDims,
-    lr: float | None = None,
-) -> Optimizer:
-    def muon_enabled(n, p):
-        if p.ndim < 2:
-            return False
-        if "lm_head" in n:
-            return False
-        if "embed_tokens" in n:
-            return False
-        return True
-
-    muon_params = []
-    expert_params = []
-    router_params = []
-    adamw_params = []
-    for n, p in named_params:
-        if p.requires_grad and muon_enabled(n, p):
-            if "mlp.experts" in n:
-                expert_params.append(p)
-            elif "mlp.router" in n:
-                router_params.append(p)
-            else:
-                muon_params.append(p)
-        elif p.requires_grad:
-            adamw_params.append(p)
-        else:
-            pass
-
-    param_groups = []
-
-    param_groups.append(
-        dict(params=muon_params, algorithm="muon", lr=lr, weight_decay=config.weight_decay, adjust_lr="rms_norm")
-    )
-    if expert_params:
-        experts_mesh_name = None
-        if parallel_dims.ep_enabled:
-            experts_mesh_name = "dp_shard_mod_ep"
-        param_groups.append(
-            dict(
-                params=expert_params,
-                algorithm="muon",
-                lr=lr,
-                weight_decay=config.weight_decay,
-                adjust_lr="rms_norm",
-                distributed_mesh_name=experts_mesh_name,
-            )
-        )
-    if router_params:
-        param_groups.append(
-            dict(
-                params=router_params,
-                algorithm="muon",
-                lr=lr,
-                weight_decay=config.weight_decay,
-                adjust_lr="rms_norm",
-            )
-        )
-
-    param_groups.append(dict(params=adamw_params, algorithm="adamw", lr=lr, weight_decay=config.weight_decay))
-
-    if parallel_dims.dp_shard_enabled or parallel_dims.cp_enabled:
-        distributed_mesh = parallel_dims.get_mesh("dp_shard_cp")
-    else:
-        distributed_mesh = parallel_dims.world_mesh
-
-    optimizer = Muon(
-        params=param_groups,
-        lr=lr,
-        mu=config.mu,
-        betas=(config.betas1, config.betas2),
-        weight_decay=config.weight_decay,
-        adjust_lr="rms_norm",
-        distributed_mesh=distributed_mesh,
-        world_mesh=parallel_dims.world_mesh,
-        fsdp_mesh_dim=1 if parallel_dims.dp_replicate_enabled else 0,
-    )
-    return optimizer
