+++ src/prime_rl/trainer/rl/train.py
@@ -50,14 +50,13 @@ from prime_rl.trainer.utils import (
     MemoryProfiler,
     Tensors,
     begin_backward,
-    bind_process_to_gpu_numa_node,
     clip_grad_norm_,
-    configure_cpu_optimizer_threads,
     filter_rl_trainer_tensor_stats_for_wandb,
     finish_backward,
     get_ckpt_disk_metrics,
     prepare_gradient_offload,
     scale_gradients_,
+    setup_cpu_optimizer_offload,
     setup_torch_distributed,
 )
 from prime_rl.trainer.world import get_world
@@ -111,10 +110,7 @@ def train(config: TrainerConfig):
     setup_torch_distributed(
         timeout=timedelta(seconds=config.dist_timeout_seconds), enable_gloo=config.model.fsdp_cpu_offload
     )
-    if config.model.optim_cpu_offload:
-        if config.model.optim_cpu_offload.numa_bind:
-            bind_process_to_gpu_numa_node()
-        configure_cpu_optimizer_threads()
+    setup_cpu_optimizer_offload(config.model.optim_cpu_offload)
     # Configurable to support ROCm/AMD GPUs where reduced precision
     # matmul corrupts softmax over large vocabularies. Override via config
     # (e.g. matmul_precision = "highest") on ROCm.
