+++ src/prime_rl/trainer/sft/train.py
@@ -41,14 +41,13 @@ from prime_rl.trainer.utils import (
     GarbageCollection,
     MemoryProfiler,
     begin_backward,
-    bind_process_to_gpu_numa_node,
-    configure_cpu_optimizer_threads,
     clip_grad_norm_,
     finish_backward,
     get_ckpt_disk_metrics,
     prepare_gradient_offload,
     print_sample,
     scale_gradients_,
+    setup_cpu_optimizer_offload,
     setup_torch_distributed,
 )
 from prime_rl.trainer.world import get_world
@@ -84,10 +83,7 @@ def train(config: SFTConfig):
     setup_torch_distributed(
         timeout=timedelta(seconds=config.dist_timeout_seconds), enable_gloo=config.model.fsdp_cpu_offload
     )
-    if config.model.optim_cpu_offload:
-        if config.model.optim_cpu_offload.numa_bind:
-            bind_process_to_gpu_numa_node()
-        configure_cpu_optimizer_threads()
+    setup_cpu_optimizer_offload(config.model.optim_cpu_offload)
     # Configurable to support ROCm/AMD GPUs where reduced precision
     # matmul corrupts softmax over large vocabularies. Override via config
     # (e.g. matmul_precision = "highest") on ROCm.
